+++ Sources/ScreenshotScribbler/Config/VerticalAlignment.swift
@@ -1,0 +1,19 @@
+//
+// Copyright © 2022 Christoph Göldner. All rights reserved.
+//
+
+import Foundation
+
+/// Vertical alignment.
+public enum VerticalAlignment: String {
+    
+    /// Vertical top alignment.
+    case top
+    
+    /// Vertical middle alignment.
+    case middle
+    
+    /// Vertical bottom alignment.
+    case bottom
+    
+}